attached_assets/qig-tokenizer/qig-tokenizer-main/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qig-tokenizer"
version = "0.1.0"
description = "QIG-native tokenizer with entropy-guided merging"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "QIG Team" }
]
keywords = ["tokenizer", "qig", "entropy", "geometric"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.15.3",
]

[project.optional-dependencies]
kernel = [
    "torch>=2.0.0",
]
redis = [
    "redis>=5.0.0",
]
postgres = [
    "psycopg2-binary>=2.9.0",
]
storage = [
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.0",
]

[project.urls]
Homepage = "https://github.com/qig/qig-tokenizer"
Documentation = "https://github.com/qig/qig-tokenizer#readme"
Repository = "https://github.com/qig/qig-tokenizer"

[tool.hatch.build.targets.wheel]
packages = ["src/qig_tokenizer"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
